refactor(RoomView): finish type colocation and drop carried-over comments
