+++ app/lib/store/definitions.ts
@@ -23,15 +23,9 @@ export type TRoomViewRoom =
 export interface IRoomStoreInitParams {
 	tmid?: string;
 	onThreadMessagesLoaded?: () => void;
-	// Per-run cancel token: a run whose signal aborts stops retrying and reports `skipped`, so a
-	// superseded run can never write over the run that replaced it.
 	signal?: AbortSignal;
 }
 
-// The distinct outcomes of one init() run. `skipped` means the run produced nothing the caller may
-// act on: either no work was attempted (no rid, an invite subscription) or the run was aborted and
-// abandoned, which can happen even after a successful load. Only `loaded` carries an unread divider
-// anchor; `failed` means every attempt was made and none succeeded.
 export type TRoomInitResult = { status: 'loaded'; lastSeen: Date | null } | { status: 'skipped' } | { status: 'failed' };
 
 export interface RoomState {
@@ -47,7 +41,6 @@ export interface RoomState {
 	canViewCannedResponse: boolean;
 	canPlaceLivechatOnHold: boolean;
 	lastMessageFromAgent: boolean;
-	// Resolves with the run's outcome; only `loaded` carries the screen's unread divider anchor.
 	init: (params?: IRoomStoreInitParams) => Promise<TRoomInitResult>;
 	join: () => void;
 	joinRoom: (requestJoinCode?: () => void) => Promise<void>;
